Completed GUI Layout
GUI is constructed and set up with apporpriate labels/fields/buttons. Functionality now needs to be added to load in Infants. This will require methods that interact with user, extract data from loaded infants, and a couple new methods in the infant class.
All TODO's are still in place. Completed items have been marked with " - check"